update: add new migration [add organization relation foreign key, add prospect type]

+++ database/migrations/2025_08_16_142901_modify_table_crm_prospect_add_prospect_type_column.php
@@ -12,7 +12,29 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('crm_prospect', function (Blueprint $table) {
-            $table->enum('prospect_type', ['donatur', 'organization', 'grab_organization'])->default('donatur')->after('nominal');
+            $table->unsignedBigInteger('donatur_id')->nullable(true)->after('id')->change();
+            
+            if (!Schema::hasColumn('crm_prospect', 'prospect_type')) {
+                $table->enum('prospect_type', ['donatur', 'organization', 'grab_organization'])->default('donatur')->after('nominal');
+            }
+
+            if (!Schema::hasColumn('crm_prospect', 'organization_id')) {
+                $table->bigInteger('organization_id')->nullable(true)->after('donatur_id');
+
+                $table->foreign('organization_id')
+                  ->references('id')
+                  ->on('organization')
+                  ->onDelete('cascade');
+            }
+
+            if (!Schema::hasColumn('crm_prospect', 'grab_organization_id')) {
+                $table->bigInteger('grab_organization_id')->nullable(true)->after('organization_id');
+
+                $table->foreign('grab_organization_id')
+                ->references('id')
+                ->on('grab_organization')
+                ->onDelete('cascade');
+            }
         });
     }
 
@@ -21,8 +43,22 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('crm_prospect', function (Blueprint $table) {
-            $table->dropColumn('prospect_type');
+         Schema::table('crm_prospect', function (Blueprint $table) {
+            if (Schema::hasColumn('crm_prospect', 'organization_id')) {
+                $table->dropForeign(['organization_id']);
+                $table->dropColumn('organization_id');
+            }
+
+            if (Schema::hasColumn('crm_prospect', 'grab_organization_id')) {
+                $table->dropForeign(['grab_organization_id']);
+                $table->dropColumn('grab_organization_id');
+            }
+
+            if (Schema::hasColumn('crm_prospect', 'prospect_type')) {
+                $table->dropColumn('prospect_type');
+            }
+
+            $table->unsignedBigInteger('donatur_id')->nullable(false)->change();
         });
     }
 };
